% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{GammaModel}
\alias{GammaModel}
\title{GammaModel}
\usage{
GammaModel(b, X)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}
}
\value{
GammaModel returns the mean function of gamma regression.
}
\description{
GammaModel is the mean function of gamma regression.
}
