% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{GaussianLogLik}
\alias{GaussianLogLik}
\title{GaussianLogLik}
\usage{
GaussianLogLik(b, X, y, sigma = 1)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}

\item{y}{outcome variable in matrix form.}

\item{sigma}{standard deviation. The default value is 1 if not specified.}
}
\value{
LogisticLogLik returns the logistic log likelihood.
}
\description{
GaussianLogLik produces gaussian probability.
}
