% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{GaussianLogPrior}
\alias{GaussianLogPrior}
\title{GaussianLogPrior}
\usage{
GaussianLogPrior(x, mu = 0, sigma = 1)
}
\arguments{
\item{x}{Regression Coefficients}

\item{mu}{Mean. The default value is 0 if not specified.}

\item{sigma}{Standard Deviation. The default value is 1 if not specified.}
}
\value{
GaussianLogPrior returns gaussian prior in log scale.
}
\description{
GaussianLogPrior is used to generate guassian prior of regression coefficients.
}
\examples{
 beta = runif(10)
 GaussianLogPrior(beta, 0, 1)

}
