% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{GaussianModel}
\alias{GaussianModel}
\title{GaussianModel}
\usage{
GaussianModel(b, X)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}
}
\value{
GaussianModel returns the mean function of gaussian regression.
}
\description{
GaussianModel is the mean function of gaussian regression.
}
