% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{LaplaceLogPrior}
\alias{LaplaceLogPrior}
\title{LaplaceLogPrior}
\usage{
LaplaceLogPrior(x, mu, b)
}
\arguments{
\item{x}{Regression Coefficients}

\item{mu}{location parameter}

\item{b}{scale parameter}
}
\value{
LaplaceLogPrior returns Laplace prior in log scale.
}
\description{
LaplaceLogPrior is used to generate Laplace prior of regression coefficients.
}
\examples{
 beta = runif(10)
 LaplaceLogPrior(beta, 0, 1)

}
