% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{LogisticLogLik}
\alias{LogisticLogLik}
\title{LogisticLogLik}
\usage{
LogisticLogLik(b, X, y)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}

\item{y}{outcome variable in matrix form.}
}
\value{
LogisticLogLik returns the logistic log likelihood.
}
\description{
LogisticLogLik produces logistic probability.
}
