% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{LogisticModel}
\alias{LogisticModel}
\title{LogisticModel}
\usage{
LogisticModel(b, X)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}
}
\value{
LogisticModel returns the log version of the mean function of logistic regression.
}
\description{
LogisticModel is the mean function (also known as sigmoid function) of logistic regression.
}
