% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{logPosteriorFunction}
\alias{logPosteriorFunction}
\title{logPosteriorFunction}
\usage{
logPosteriorFunction(b, X, y)
}
\arguments{
\item{b}{Beta Coefficients in matrix form.}

\item{X}{Sets of predictors in matrix form.}

\item{y}{outcome variable in matrix form.}
}
\value{
logPosteriorFunction returns un-normalized posterior pdf.
}
\description{
logPosteriorFunction produces un-normalized posterior pdf
}
