% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC.beta.generate}
\alias{MCMC.beta.generate}
\title{MCMC.beta.generate}
\usage{
MCMC.beta.generate(model, prior, likelihood, posterior, X, y)
}
\arguments{
\item{model}{Regression model.}

\item{prior}{Prior functions.}

\item{likelihood}{Log likelihood function.}

\item{posterior}{Posterior function.}

\item{X}{Sets of predictors.}

\item{y}{Outcome.}
}
\value{
MCMC.beta.generate returns the samples of beta coefficients.
}
\description{
MCMC.beta.generate is used to generate samples of beta coefficients for top half features selected using PCA.
}
