% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC.prediction}
\alias{MCMC.prediction}
\title{MCMC.prediction}
\usage{
MCMC.prediction(model, likelihood, sample, X, selected.clusters)
}
\arguments{
\item{model}{Regression model.}

\item{likelihood}{Log likelihood function.}

\item{sample}{samples of beta coefficients genereated with MCMC.beta.generate.}

\item{X}{Sets of predictors.}

\item{selected.clusters}{Selected cluster labels as a vector.}
}
\value{
MCMC.prediction returns predicted outcome with samples of beta coefficients.
}
\description{
MCMC.prediction is used to generate predictive outcomes.
}
