% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pSVN.R
\name{pSVN.beta.generate}
\alias{pSVN.beta.generate}
\title{pSVN.beta.generate}
\usage{
pSVN.beta.generate(model, prior, likelihood, X, y)
}
\arguments{
\item{model}{Regression model.}

\item{prior}{Prior functions.}

\item{likelihood}{Log likelihood function.}

\item{X}{Sets of predictors.}

\item{y}{Outcome.}
}
\value{
pSVN.beta.generate returns the samples of beta coefficients.
}
\description{
pSVN.beta.generate is used to generate samples of beta coefficients for top half features selected using PCA.
}
