% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGD.R
\name{SVGD.beta.generate}
\alias{SVGD.beta.generate}
\title{SVGD.beta.generate}
\usage{
SVGD.beta.generate(X, y)
}
\arguments{
\item{X}{Sets of predictors.}

\item{y}{Outcome.}
}
\value{
MCMC.beta.generate returns the samples of beta coefficients.
}
\description{
SVGD.beta.generate is used to generate samples of beta coefficients for top 1/2 features selected using PCA.
}
