% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFns.R
\name{GaussianLogPrior}
\alias{GaussianLogPrior}
\title{GaussianLogPrior}
\usage{
GaussianLogPrior(x, mu, sigma)
}
\arguments{
\item{x}{Regression Coefficients}

\item{mu}{Mean}

\item{sigma}{Standard Deviation}
}
\value{
GaussianLogPrior returns gaussian prior in log scale.
}
\description{
GaussianLogPrior is used to generate guassian prior of regression coefficients.
}
\examples{
 beta = runif(10)
 GaussianLogPrior(beta, 0, 1)

}
